+++ release/githubio/js/main_uncompressed.js
@@ -7009,7 +7009,7 @@
     var Main_stringVersion_Min = '.267';
     var Main_version_java = 50; //Always update (+1 to current value) Main_version_java after update Main_stringVersion_Min or a major update of the apk is released
     var Main_minversion = 'October 21 2020';
-    var Main_version_web = 107; //Always update (+1 to current value) Main_version_web after update Main_minversion or a major update of the web part of the app
+    var Main_version_web = 108; //Always update (+1 to current value) Main_version_web after update Main_minversion or a major update of the web part of the app
     var Main_versionTag = Main_stringVersion + Main_stringVersion_Min + '-' + Main_minversion;
 
     var Main_cursorYAddFocus = -1;
@@ -7596,6 +7596,7 @@
                 title: "Web Version October 21 2020",
                 changes: [
                     "Add a progress indicator to already watched VOD/Clip",
+                    "Save a clip as watched when the previewed clip ends (of course if history is enabled for clips)"
                 ]
             },
             {
@@ -8710,8 +8711,8 @@
             ArrayPos.views = Data[13];
 
         } else {
-            //Limit size to 1500
-            if (Main_values_History_data[AddUser_UsernameArray[0].id][type].length > 1499) {
+            //Limit size to 2000
+            if (Main_values_History_data[AddUser_UsernameArray[0].id][type].length > 1999) {
 
                 //Sort by oldest first to delete the oldest
                 Main_values_History_data[AddUser_UsernameArray[0].id][type].sort(
@@ -15558,6 +15559,18 @@
 
             if (ScreenObj[Main_values.Main_Go].screenType === 2 && Settings_Obj_default('auto_clip_preview')) {
 
+                var id = ScreenObj[Main_values.Main_Go].posY + '_' + ScreenObj[Main_values.Main_Go].posX;
+
+                if (ScreenObj[Main_HistoryClip].histPosX[1]) {
+                    //Save as we have watched it all
+
+                    var data = JSON.parse(Main_getElementById(ScreenObj[Main_values.Main_Go].ids[3] + id).getAttribute(Main_DataAttribute));
+
+                    Main_getElementById(ScreenObj[Main_values.Main_Go].ids[7] + id).style.width = '100%';
+
+                    Main_history_UpdateVodClip(data[7], data[1], 'clip');
+                }
+
                 if (PlayClip_getIdNext(1, 0)) {
                     //Use OSInterface_keyEvent to prevent odd screen scroll visual behavior
                     OSInterface_keyEvent(3, 0);
@@ -15572,7 +15585,7 @@
                     Play_CheckIfIsLiveCleanEnd();
 
                     Main_RemoveClass(
-                        ScreenObj[Main_values.Main_Go].ids[1] + ScreenObj[Main_values.Main_Go].posY + '_' + ScreenObj[Main_values.Main_Go].posX,
+                        ScreenObj[Main_values.Main_Go].ids[1] + id,
                         'opacity_zero'
                     );
                 }