+++ apk/Changelog.md
@@ -13,6 +13,7 @@ Web Version October 21 2020
 ==============
 
 * Add a progress indicator to already watched VOD/Clip
+* Save a clip as watched when the previewed clip ends (of course if history is enabled for clips)
 
 Web Version October 20 2020
 ==============